Move UpgradeResourceState code closer to assist refactoring
